requirements.txt
flask
flask-restx
pyyaml